+++ blaze/__init__.py
@@ -58,6 +58,7 @@ from .expr import (
 )
 from .expr.arrays import (tensordot, transpose)
 from .expr.functions import *
+from .expr.literal import data
 from .index import create_index
 from .interactive import *
 from .compute.pmap import set_default_pmap
