+++ blaze/index.py
@@ -1,7 +1,8 @@
 from .dispatch import dispatch
 from .compatibility import basestring
-from blaze.interactive import _Data
-from blaze.interactive import data as bz_data
+from blaze.expr.literal import Data
+from blaze.expr.literal import data as bz_data
+
 
 @dispatch(object, (basestring, list, tuple))
 def create_index(t, column_name_or_names, name=None):
@@ -34,7 +35,7 @@ def create_index(t, column_name_or_names, name=None):
     raise NotImplementedError("create_index not implemented for type %r" %
                               type(t).__name__)
 
-@dispatch(_Data, (basestring, list, tuple))
+@dispatch(Data, (basestring, list, tuple))
 def create_index(dta, column_name_or_names, name=None, **kwargs):
     return create_index(dta.data, column_name_or_names, name=name, **kwargs)
